build: updates react-output-target version
- updates `react-output-target` version for component-library and component-library-react.
- swaps content for SSR and Client pages in nextJS app
